+++ packages/sdk/src/plt/v1/Governance.ts
@@ -14,6 +14,13 @@ import {
     createTokenGovernancePayload,
 } from './types.js';
 
+type SupplyUpdateOptions = {
+    /** Whether to automatically scale a token amount to the correct number of decimals as the token */
+    autoScale?: boolean;
+    /** Whether to validate the payload executing it */
+    validate?: boolean;
+};
+
 /**
  * Mints a specified amount of tokens.
  *
@@ -22,6 +29,7 @@ import {
  * @param {TokenAmount.Type | TokenAmount.Type[]} amounts - The amount(s) of tokens to mint.
  * @param {AccountSigner} signer - The signer responsible for signing the transaction.
  * @param {TransactionExpiry.Type} [expiry=TransactionExpiry.futureMinutes(5)] - The expiry time for the transaction.
+ * @param {SupplyUpdateOptions} [opts={ autoScale: true, validate: true }] - Options for supply update operations.
  *
  * @returns A promise that resolves to the transaction hash.
  * @throws {InvalidTokenAmountError} If the token amount is not compatible with the token.
@@ -32,14 +40,21 @@ export async function mint(
     sender: AccountAddress.Type,
     amounts: TokenAmount.Type | TokenAmount.Type[],
     signer: AccountSigner,
-    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5)
+    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5),
+    opts: SupplyUpdateOptions = { autoScale: true, validate: true }
 ): Promise<TransactionHash.Type> {
-    const amountsList = [amounts].flat();
-    // TODO: re-enable validation when it's covered by unit tests
-    // amountsList.forEach((amount) => validateAmount(token, amount));
+    let amountsList = [amounts].flat();
+    if (opts.autoScale) {
+        amountsList = amountsList.map((amount) => scaleAmount(token, amount));
+    }
+
+    if (opts.validate) {
+        // TODO: re-enable validation when it's covered by unit tests
+        // amountsList.forEach((amount) => validateAmount(token, amount));
+    }
 
     const ops: TokenMintOperation[] = amountsList.map((amount) => ({
-        [TokenOperationType.Mint]: { amount: scaleAmount(token, amount) },
+        [TokenOperationType.Mint]: { amount },
     }));
     return batchOperations(token, sender, ops, signer, expiry);
 }
@@ -52,6 +67,7 @@ export async function mint(
  * @param {TokenAmount.Type | TokenAmount.Type[]} amounts - The amount(s) of tokens to burn.
  * @param {AccountSigner} signer - The signer responsible for signing the transaction.
  * @param {TransactionExpiry.Type} [expiry=TransactionExpiry.futureMinutes(5)] - The expiry time for the transaction.
+ * @param {SupplyUpdateOptions} [opts={ autoScale: true, validate: true }] - Options for supply update operations.
  *
  * @returns A promise that resolves to the transaction hash.
  * @throws {InvalidTokenAmountError} If the token amount is not compatible with the token.
@@ -62,17 +78,29 @@ export async function burn(
     sender: AccountAddress.Type,
     amounts: TokenAmount.Type | TokenAmount.Type[],
     signer: AccountSigner,
-    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5)
+    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5),
+    opts: SupplyUpdateOptions = { autoScale: true, validate: true }
 ): Promise<TransactionHash.Type> {
-    const amountsList = [amounts].flat();
-    // TODO: re-enable validation when it's covered by unit tests
-    // amountsList.forEach((amount) => validateAmount(token, amount));
+    let amountsList = [amounts].flat();
+    if (opts.autoScale) {
+        amountsList = amountsList.map((amount) => scaleAmount(token, amount));
+    }
+
+    if (opts.validate) {
+        // TODO: re-enable validation when it's covered by unit tests
+        // amountsList.forEach((amount) => validateAmount(token, amount));
+    }
 
     const ops: TokenBurnOperation[] = amountsList.map((amount) => ({
-        [TokenOperationType.Burn]: { amount: scaleAmount(token, amount) },
+        [TokenOperationType.Burn]: { amount },
     }));
     return batchOperations(token, sender, ops, signer, expiry);
 }
+
+type UpdateListOptions = {
+    /** Whether to validate the payload executing it */
+    validate?: boolean;
+};
 
 /**
  * Adds an account to the allow list of a token.
@@ -82,6 +110,7 @@ export async function burn(
  * @param {AccountAddress.Type | AccountAddress.Type[]} targets - The account address(es) to be added to the list.
  * @param {AccountSigner} signer - The signer responsible for signing the transaction.
  * @param {TransactionExpiry.Type} [expiry=TransactionExpiry.futureMinutes(5)] - The expiry time for the transaction.
+ * @param {UpdateListOptions} [opts={ validate: true }] - Options for updating the allow/deny list.
  *
  * @returns A promise that resolves to the transaction hash.
  * @throws {UnauthorizedGovernanceOperationError} If the sender is not the token issuer.
@@ -91,12 +120,13 @@ export async function addAllowList(
     sender: AccountAddress.Type,
     targets: AccountAddress.Type | AccountAddress.Type[],
     signer: AccountSigner,
-    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5)
+    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5),
+    { validate }: UpdateListOptions = { validate: true }
 ): Promise<TransactionHash.Type> {
     const ops: TokenAddAllowListOperation[] = [targets]
         .flat()
         .map((target) => ({ [TokenOperationType.AddAllowList]: { target } }));
-    return batchOperations(token, sender, ops, signer, expiry);
+    return batchOperations(token, sender, ops, signer, expiry, validate);
 }
 
 /**
@@ -107,6 +137,7 @@ export async function addAllowList(
  * @param {AccountAddress.Type | AccountAddress.Type[]} targets - The account address(es) to be added to the list.
  * @param {AccountSigner} signer - The signer responsible for signing the transaction.
  * @param {TransactionExpiry.Type} [expiry=TransactionExpiry.futureMinutes(5)] - The expiry time for the transaction.
+ * @param {UpdateListOptions} [opts={ validate: true }] - Options for updating the allow/deny list.
  *
  * @returns A promise that resolves to the transaction hash.
  * @throws {UnauthorizedGovernanceOperationError} If the sender is not the token issuer.
@@ -116,12 +147,13 @@ export async function removeAllowList(
     sender: AccountAddress.Type,
     targets: AccountAddress.Type | AccountAddress.Type[],
     signer: AccountSigner,
-    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5)
+    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5),
+    { validate }: UpdateListOptions = { validate: true }
 ): Promise<TransactionHash.Type> {
     const ops: TokenRemoveAllowListOperation[] = [targets]
         .flat()
         .map((target) => ({ [TokenOperationType.RemoveAllowList]: { target } }));
-    return batchOperations(token, sender, ops, signer, expiry);
+    return batchOperations(token, sender, ops, signer, expiry, validate);
 }
 
 /**
@@ -132,6 +164,7 @@ export async function removeAllowList(
  * @param {AccountAddress.Type | AccountAddress.Type[]} targets - The account address(es) to be added to the list.
  * @param {AccountSigner} signer - The signer responsible for signing the transaction.
  * @param {TransactionExpiry.Type} [expiry=TransactionExpiry.futureMinutes(5)] - The expiry time for the transaction.
+ * @param {UpdateListOptions} [opts={ validate: true }] - Options for updating the allow/deny list.
  *
  * @returns A promise that resolves to the transaction hash.
  * @throws {UnauthorizedGovernanceOperationError} If the sender is not the token issuer.
@@ -141,12 +174,13 @@ export async function addDenyList(
     sender: AccountAddress.Type,
     targets: AccountAddress.Type | AccountAddress.Type[],
     signer: AccountSigner,
-    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5)
+    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5),
+    { validate }: UpdateListOptions = { validate: true }
 ): Promise<TransactionHash.Type> {
     const ops: TokenAddDenyListOperation[] = [targets]
         .flat()
         .map((target) => ({ [TokenOperationType.AddDenyList]: { target } }));
-    return batchOperations(token, sender, ops, signer, expiry);
+    return batchOperations(token, sender, ops, signer, expiry, validate);
 }
 
 /**
@@ -157,6 +191,7 @@ export async function addDenyList(
  * @param {AccountAddress.Type | AccountAddress.Type[]} targets - The account address(es) to be added to the list.
  * @param {AccountSigner} signer - The signer responsible for signing the transaction.
  * @param {TransactionExpiry.Type} [expiry=TransactionExpiry.futureMinutes(5)] - The expiry time for the transaction.
+ * @param {UpdateListOptions} [opts={ validate: true }] - Options for updating the allow/deny list.
  *
  * @returns A promise that resolves to the transaction hash.
  * @throws {UnauthorizedGovernanceOperationError} If the sender is not the token issuer.
@@ -166,12 +201,13 @@ export async function removeDenyList(
     sender: AccountAddress.Type,
     targets: AccountAddress.Type | AccountAddress.Type[],
     signer: AccountSigner,
-    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5)
+    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5),
+    { validate }: UpdateListOptions = { validate: true }
 ): Promise<TransactionHash.Type> {
     const ops: TokenRemoveDenyListOperation[] = [targets]
         .flat()
         .map((target) => ({ [TokenOperationType.RemoveDenyList]: { target } }));
-    return batchOperations(token, sender, ops, signer, expiry);
+    return batchOperations(token, sender, ops, signer, expiry, validate);
 }
 
 /**
@@ -182,6 +218,7 @@ export async function removeDenyList(
  * @param {TokenGovernanceOperation[]} operations - An array of governance operations to execute.
  * @param {AccountSigner} signer - The signer responsible for signing the transaction.
  * @param {TransactionExpiry.Type} [expiry=TransactionExpiry.futureMinutes(5)] - The expiry time for the transaction.
+ * @param {boolean} [validate=true] - Whether to validate the operations before executing.
  *
  * @returns A promise that resolves to the transaction hash.
  * @throws {InvalidTokenAmountError} If a token amount is not compatible with the token.
@@ -192,8 +229,9 @@ export async function batchOperations(
     sender: AccountAddress.Type,
     operations: TokenGovernanceOperation[],
     signer: AccountSigner,
-    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5)
+    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5),
+    validate: boolean = true
 ): Promise<TransactionHash.Type> {
     const payload = createTokenGovernancePayload(token.info.id, operations);
-    return governanceTransaction(token, sender, payload, signer, expiry);
+    return governanceTransaction(token, sender, payload, signer, expiry, validate);
 }