+++ packages/sdk/CHANGELOG.md
@@ -13,6 +13,8 @@
   the parameter/return values are internal-only.
 - Added `TokenHolderPayload` and `TokenGovernancePayload` to `AccountTransactionPayload` union type.
 - Added reject reasons related to PLT transactions to `RejectReason` union type.
+- `CcdAmount.fromDecimal` no longer supports creation from a string with comma used as the decimal separator, e.g.
+  "10,123".
 
 ### Added
 
@@ -42,8 +44,9 @@
   any previously registered decoder for the corresponding CBOR tag.
 - `TokenHolderHandler` and `TokenGovernanceHandler`, which are also accessible by passing the corresponding
   `TransactionType` to `getAccountTransactionHandler`.
-- function `parseSimpleWallet` which parses a `SimpleWalletFormat` (also a subset of `GenesisFormat`), which can be used
+- Function `parseSimpleWallet` which parses a `SimpleWalletFormat` (also a subset of `GenesisFormat`), which can be used
   with `buildAccountSigner`
+- A new optional field `createPlt` to `AuthorizationsV1` which exposes the access structure for PLT creation.
 
 ## 9.1.1